@@ -3,7 +3,8 @@ package converter
 import (
 	"strings"
 
-	"github.com/golang/protobuf/protoc-gen-go/descriptor"
+	"github.com/sirupsen/logrus"
+	"google.golang.org/protobuf/types/descriptorpb"
 )
 
 // ProtoPackage describes a package of Protobuf, which is an container of message types.
@@ -11,57 +12,61 @@ type ProtoPackage struct {
 	name     string
 	parent   *ProtoPackage
 	children map[string]*ProtoPackage
-	types    map[string]*descriptor.DescriptorProto
+	types    map[string]*descriptorpb.DescriptorProto
 }
 
-func (c *Converter) lookupType(pkg *ProtoPackage, name string) (*descriptor.DescriptorProto, string, bool) {
+func (c *Converter) lookupType(pkg *ProtoPackage, name string) (*descriptorpb.DescriptorProto, string, bool) {
 	if strings.HasPrefix(name, ".") {
 		return c.relativelyLookupType(globalPkg, name[1:len(name)])
 	}
 
-	for ; pkg != nil; pkg = pkg.parent {
+	for pkg != nil {
 		if desc, pkgName, ok := c.relativelyLookupType(pkg, name); ok {
 			return desc, pkgName, ok
 		}
+
+		pkg = pkg.parent
 	}
 	return nil, "", false
 }
 
-func (c *Converter) relativelyLookupType(pkg *ProtoPackage, name string) (*descriptor.DescriptorProto, string, bool) {
-	components := strings.SplitN(name, ".", 2)
-	switch len(components) {
-	case 0:
-		c.logger.Debug("empty message name")
-		return nil, "", false
-	case 1:
-		found, ok := pkg.types[components[0]]
+func (c *Converter) relativelyLookupType(pkg *ProtoPackage, name string) (*descriptorpb.DescriptorProto, string, bool) {
+	head, tail, _ := strings.Cut(name, ".")
+
+	if tail == "" {
+		found, ok := pkg.types[head]
 		return found, pkg.name, ok
-	case 2:
-		c.logger.Tracef("Looking for %s in %s at %s (%v)", components[1], components[0], pkg.name, pkg)
-		if child, ok := pkg.children[components[0]]; ok {
-			found, pkgName, ok := c.relativelyLookupType(child, components[1])
-			return found, pkgName, ok
-		}
-		if msg, ok := pkg.types[components[0]]; ok {
-			found, ok := c.relativelyLookupNestedType(msg, components[1])
-			return found, pkg.name, ok
-		}
-		c.logger.WithField("component", components[0]).WithField("package_name", pkg.name).Info("No such package nor message in package")
-		return nil, "", false
-	default:
-		c.logger.Error("Failed to lookup type")
-		return nil, "", false
 	}
+
+	c.Logger.Tracef("Looking for %s in %s at %s (%v)", tail, head, pkg.name, pkg)
+
+	if child := pkg.children[head]; child != nil {
+		found, pkgName, ok := c.relativelyLookupType(child, tail)
+		return found, pkgName, ok
+	}
+
+	if msg := pkg.types[head]; msg != nil {
+		found, ok := c.relativelyLookupNestedType(msg, tail)
+		return found, pkg.name, ok
+	}
+
+	c.Logger.WithFields(logrus.Fields{
+		"component": head,
+		"package_name": pkg.name,
+	}).Info("No such package nor message in package")
+
+	return nil, "", false
 }
 
 func (c *Converter) relativelyLookupPackage(pkg *ProtoPackage, name string) (*ProtoPackage, bool) {
-	components := strings.Split(name, ".")
-	for _, c := range components {
-		var ok bool
-		pkg, ok = pkg.children[c]
-		if !ok {
+	for _, c := range strings.Split(name, ".") {
+		next := pkg.children[c]
+		if next == nil {
 			return nil, false
 		}
+
+		pkg = next
 	}
+
 	return pkg, true
 }
