@@ -2,18 +2,17 @@ package converter
 
 import (
 	"bytes"
-	"fmt"
 	"os"
 	"os/exec"
 	"strings"
 	"testing"
 
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/protobuf/protoc-gen-go/descriptor"
-	plugin "github.com/golang/protobuf/protoc-gen-go/plugin"
+	"github.com/google/go-cmp/cmp"
+	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/descriptorpb"
+	"google.golang.org/protobuf/types/pluginpb"
 	"github.com/sirupsen/logrus"
 	"github.com/sixt/protoc-gen-jsonschema/internal/converter/testdata"
-	"github.com/stretchr/testify/assert"
 )
 
 var (
@@ -22,11 +21,11 @@ var (
 )
 
 type sampleProto struct {
-	AllowNullValues           bool
 	ExpectedJSONSchema        []string
 	FilesToGenerate           []string
 	ProtoFileName             string
 	UseProtoAndJSONFieldNames bool
+	DisallowAdditionalProperties bool
 }
 
 func TestGenerateJsonSchema(t *testing.T) {
@@ -35,24 +34,27 @@ func TestGenerateJsonSchema(t *testing.T) {
 	configureSampleProtos()
 
 	// Convert the protos, compare the results against the expected JSON-Schemas:
-	testConvertSampleProto(t, sampleProtos["Comments"])
-	testConvertSampleProto(t, sampleProtos["ArrayOfMessages"])
-	testConvertSampleProto(t, sampleProtos["ArrayOfObjects"])
-	testConvertSampleProto(t, sampleProtos["ArrayOfPrimitives"])
-	testConvertSampleProto(t, sampleProtos["ArrayOfPrimitivesDouble"])
-	testConvertSampleProto(t, sampleProtos["EnumCeption"])
-	testConvertSampleProto(t, sampleProtos["ImportedEnum"])
-	testConvertSampleProto(t, sampleProtos["NestedMessage"])
-	testConvertSampleProto(t, sampleProtos["NestedObject"])
-	testConvertSampleProto(t, sampleProtos["PayloadMessage"])
-	testConvertSampleProto(t, sampleProtos["SeveralEnums"])
-	testConvertSampleProto(t, sampleProtos["SeveralMessages"])
-	testConvertSampleProto(t, sampleProtos["ArrayOfEnums"])
-	testConvertSampleProto(t, sampleProtos["Maps"])
-	testConvertSampleProto(t, sampleProtos["WellKnown"])
+	testConvertSampleProto(t, "Comments")
+	testConvertSampleProto(t, "ArrayOfMessages")
+	testConvertSampleProto(t, "ArrayOfObjects")
+	testConvertSampleProto(t, "ArrayOfPrimitives")
+	testConvertSampleProto(t, "ArrayOfPrimitivesDouble")
+	testConvertSampleProto(t, "Enumception")
+	testConvertSampleProto(t, "ImportedEnum")
+	testConvertSampleProto(t, "NestedMessage")
+	testConvertSampleProto(t, "NestedObject")
+	testConvertSampleProto(t, "PayloadMessage")
+	testConvertSampleProto(t, "SeveralEnums")
+	testConvertSampleProto(t, "SeveralMessages")
+	testConvertSampleProto(t, "ArrayOfEnums")
+	testConvertSampleProto(t, "Maps")
+	testConvertSampleProto(t, "WellKnown")
 }
 
-func testConvertSampleProto(t *testing.T, sampleProto sampleProto) {
+func testConvertSampleProto(t *testing.T, name string) {
+
+	sampleProto := sampleProtos[name]
+	t.Log("running test:", name)
 
 	// Make a Logrus logger:
 	logger := logrus.New()
@@ -60,38 +62,44 @@ func testConvertSampleProto(t *testing.T, sampleProto sampleProto) {
 	logger.SetOutput(os.Stderr)
 
 	// Use the logger to make a Converter:
-	protoConverter := New(logger)
-	protoConverter.AllowNullValues = sampleProto.AllowNullValues
+	protoConverter := &Converter{
+		Logger: logger,
+	}
 	protoConverter.UseProtoAndJSONFieldnames = sampleProto.UseProtoAndJSONFieldNames
+	protoConverter.DisallowAdditionalProperties = sampleProto.DisallowAdditionalProperties
 
 	// Open the sample proto file:
-	sampleProtoFileName := fmt.Sprintf("%v/%v", sampleProtoDirectory, sampleProto.ProtoFileName)
+	sampleProtoFileName := sampleProtoDirectory + "/" + sampleProto.ProtoFileName
 	fileDescriptorSet := mustReadProtoFiles(t, sampleProtoDirectory, sampleProto.ProtoFileName)
 
 	// Prepare a request:
-	codeGeneratorRequest := plugin.CodeGeneratorRequest{
+	codeGeneratorRequest := &pluginpb.CodeGeneratorRequest{
 		FileToGenerate: sampleProto.FilesToGenerate,
 		ProtoFile:      fileDescriptorSet.GetFile(),
 	}
 
 	// Perform the conversion:
-	response, err := protoConverter.convert(&codeGeneratorRequest)
-	assert.NoError(t, err, "Unable to convert sample proto file (%v)", sampleProtoFileName)
-	assert.Equal(t, len(sampleProto.ExpectedJSONSchema), len(response.File), "Incorrect number of JSON-Schema files returned for sample proto file (%v)", sampleProtoFileName)
-	if len(sampleProto.ExpectedJSONSchema) != len(response.File) {
-		t.Fail()
-	} else {
-		for responseFileIndex, responseFile := range response.File {
-			assert.Equal(t, sampleProto.ExpectedJSONSchema[responseFileIndex], *responseFile.Content, "Incorrect JSON-Schema returned for sample proto file (%v)", sampleProtoFileName)
-		}
+	response, err := protoConverter.convert(codeGeneratorRequest)
+	if err != nil {
+		t.Fatal(err)
 	}
 
+	if len(response.File) != len(sampleProto.ExpectedJSONSchema) {
+		t.Fatal("Incorrect number of JSON-Schema files returned for sample proto file:", sampleProtoFileName)
+	}
+
+	for i, file := range response.File {
+		want := sampleProto.ExpectedJSONSchema[i]
+
+		if diff := cmp.Diff(file.GetContent(), want); diff != "" {
+			t.Errorf("differences: %s\n%s", file.GetName(), diff)
+		}
+	}
 }
 
 func configureSampleProtos() {
 	// ArrayOfMessages:
 	sampleProtos["ArrayOfMessages"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.PayloadMessage, testdata.ArrayOfMessages},
 		FilesToGenerate:    []string{"ArrayOfMessages.proto", "PayloadMessage.proto"},
 		ProtoFileName:      "ArrayOfMessages.proto",
@@ -99,7 +107,6 @@ func configureSampleProtos() {
 
 	// ArrayOfObjects:
 	sampleProtos["ArrayOfObjects"] = sampleProto{
-		AllowNullValues:    true,
 		ExpectedJSONSchema: []string{testdata.ArrayOfObjects},
 		FilesToGenerate:    []string{"ArrayOfObjects.proto"},
 		ProtoFileName:      "ArrayOfObjects.proto",
@@ -107,7 +114,6 @@ func configureSampleProtos() {
 
 	// ArrayOfPrimitives:
 	sampleProtos["ArrayOfPrimitives"] = sampleProto{
-		AllowNullValues:    true,
 		ExpectedJSONSchema: []string{testdata.ArrayOfPrimitives},
 		FilesToGenerate:    []string{"ArrayOfPrimitives.proto"},
 		ProtoFileName:      "ArrayOfPrimitives.proto",
@@ -115,24 +121,22 @@ func configureSampleProtos() {
 
 	// ArrayOfPrimitives:
 	sampleProtos["ArrayOfPrimitivesDouble"] = sampleProto{
-		AllowNullValues:           true,
 		ExpectedJSONSchema:        []string{testdata.ArrayOfPrimitivesDouble},
 		FilesToGenerate:           []string{"ArrayOfPrimitives.proto"},
 		ProtoFileName:             "ArrayOfPrimitives.proto",
+
 		UseProtoAndJSONFieldNames: true,
 	}
 
-	// EnumCeption:
-	sampleProtos["EnumCeption"] = sampleProto{
-		AllowNullValues:    false,
-		ExpectedJSONSchema: []string{testdata.PayloadMessage, testdata.ImportedEnum, testdata.EnumCeption},
+	// Enumception:
+	sampleProtos["Enumception"] = sampleProto{
+		ExpectedJSONSchema: []string{testdata.PayloadMessage, testdata.ImportedEnum, testdata.Enumception},
 		FilesToGenerate:    []string{"Enumception.proto", "PayloadMessage.proto", "ImportedEnum.proto"},
 		ProtoFileName:      "Enumception.proto",
 	}
 
 	// ImportedEnum:
 	sampleProtos["ImportedEnum"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.ImportedEnum},
 		FilesToGenerate:    []string{"ImportedEnum.proto"},
 		ProtoFileName:      "ImportedEnum.proto",
@@ -140,15 +144,14 @@ func configureSampleProtos() {
 
 	// NestedMessage:
 	sampleProtos["NestedMessage"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.PayloadMessage, testdata.NestedMessage},
 		FilesToGenerate:    []string{"NestedMessage.proto", "PayloadMessage.proto"},
 		ProtoFileName:      "NestedMessage.proto",
+
 	}
 
 	// NestedObject:
 	sampleProtos["NestedObject"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.NestedObject},
 		FilesToGenerate:    []string{"NestedObject.proto"},
 		ProtoFileName:      "NestedObject.proto",
@@ -156,7 +159,6 @@ func configureSampleProtos() {
 
 	// PayloadMessage:
 	sampleProtos["PayloadMessage"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.PayloadMessage},
 		FilesToGenerate:    []string{"PayloadMessage.proto"},
 		ProtoFileName:      "PayloadMessage.proto",
@@ -164,7 +166,6 @@ func configureSampleProtos() {
 
 	// SeveralEnums:
 	sampleProtos["SeveralEnums"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.FirstEnum, testdata.SecondEnum},
 		FilesToGenerate:    []string{"SeveralEnums.proto"},
 		ProtoFileName:      "SeveralEnums.proto",
@@ -172,7 +173,6 @@ func configureSampleProtos() {
 
 	// SeveralMessages:
 	sampleProtos["SeveralMessages"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.FirstMessage, testdata.SecondMessage},
 		FilesToGenerate:    []string{"SeveralMessages.proto"},
 		ProtoFileName:      "SeveralMessages.proto",
@@ -180,7 +180,6 @@ func configureSampleProtos() {
 
 	// ArrayOfEnums:
 	sampleProtos["ArrayOfEnums"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.ArrayOfEnums},
 		FilesToGenerate:    []string{"ArrayOfEnums.proto"},
 		ProtoFileName:      "ArrayOfEnums.proto",
@@ -188,7 +187,6 @@ func configureSampleProtos() {
 
 	// Maps:
 	sampleProtos["Maps"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.Maps},
 		FilesToGenerate:    []string{"Maps.proto"},
 		ProtoFileName:      "Maps.proto",
@@ -196,7 +194,6 @@ func configureSampleProtos() {
 
 	// Comments:
 	sampleProtos["Comments"] = sampleProto{
-		AllowNullValues:    false,
 		ExpectedJSONSchema: []string{testdata.MessageWithComments},
 		FilesToGenerate:    []string{"MessageWithComments.proto"},
 		ProtoFileName:      "MessageWithComments.proto",
@@ -211,33 +208,36 @@ func configureSampleProtos() {
 
 // Load the specified .proto files into a FileDescriptorSet. Any errors in loading/parsing will
 // immediately fail the test.
-func mustReadProtoFiles(t *testing.T, includePath string, filenames ...string) *descriptor.FileDescriptorSet {
+func mustReadProtoFiles(t *testing.T, includePath string, filenames ...string) *descriptorpb.FileDescriptorSet {
 	protocBinary, err := exec.LookPath("protoc")
 	if err != nil {
-		t.Fatalf("Can't find 'protoc' binary in $PATH: %s", err.Error())
+		t.Fatal("Can't find 'protoc' binary in $PATH:", err)
 	}
 
-	// Use protoc to output descriptor info for the specified .proto files.
-	var args []string
-	args = append(args, "--descriptor_set_out=/dev/stdout")
-	args = append(args, "--include_source_info")
-	args = append(args, "--include_imports")
-	args = append(args, "--proto_path="+includePath)
+	// Use protoc to output descriptorpb info for the specified .proto files.
+	args := []string{
+		"--descriptor_set_out=/dev/stdout",
+		"--include_source_info",
+		"--include_imports",
+		"--proto_path="+includePath,
+	}
 	args = append(args, filenames...)
+
 	cmd := exec.Command(protocBinary, args...)
 	stdoutBuf := bytes.Buffer{}
 	stderrBuf := bytes.Buffer{}
 	cmd.Stdout = &stdoutBuf
 	cmd.Stderr = &stderrBuf
+
 	err = cmd.Run()
 	if err != nil {
-		t.Fatalf("failed to load descriptor set (%s): %s: %s",
-			strings.Join(cmd.Args, " "), err.Error(), stderrBuf.String())
+		t.Fatalf("failed to load descriptorpb set (%s): %s: %s", strings.Join(cmd.Args, " "), err, stderrBuf.String())
 	}
-	fds := &descriptor.FileDescriptorSet{}
-	err = proto.Unmarshal(stdoutBuf.Bytes(), fds)
-	if err != nil {
-		t.Fatalf("failed to parse protoc output as FileDescriptorSet: %s", err.Error())
+
+	fds := new(descriptorpb.FileDescriptorSet)
+	if err := proto.Unmarshal(stdoutBuf.Bytes(), fds); err != nil {
+		t.Fatal("failed to parse protoc output as FileDescriptorSet:", err)
 	}
+
 	return fds
 }
