@@ -1,12 +1,12 @@
 package converter
 
 import (
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/protobuf/protoc-gen-go/descriptor"
+	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/descriptorpb"
 )
 
 // Protobuf tag values for relevant message fields. Full list here:
-//   https://github.com/protocolbuffers/protobuf/blob/master/src/google/protobuf/descriptor.proto
+//   https://github.com/protocolbuffers/protobuf/blob/master/src/google/protobuf/descriptorpb.proto
 const (
 	tag_FileDescriptor_messageType int32 = 4
 	tag_FileDescriptor_enumType    int32 = 5
@@ -18,30 +18,30 @@ const (
 )
 
 type sourceCodeInfo struct {
-	lookup map[proto.Message]*descriptor.SourceCodeInfo_Location
+	lookup map[proto.Message]*descriptorpb.SourceCodeInfo_Location
 }
 
-func (s sourceCodeInfo) GetMessage(m *descriptor.DescriptorProto) *descriptor.SourceCodeInfo_Location {
+func (s sourceCodeInfo) GetMessage(m *descriptorpb.DescriptorProto) *descriptorpb.SourceCodeInfo_Location {
 	return s.lookup[m]
 }
 
-func (s sourceCodeInfo) GetField(f *descriptor.FieldDescriptorProto) *descriptor.SourceCodeInfo_Location {
+func (s sourceCodeInfo) GetField(f *descriptorpb.FieldDescriptorProto) *descriptorpb.SourceCodeInfo_Location {
 	return s.lookup[f]
 }
 
-func (s sourceCodeInfo) GetEnum(e *descriptor.EnumDescriptorProto) *descriptor.SourceCodeInfo_Location {
+func (s sourceCodeInfo) GetEnum(e *descriptorpb.EnumDescriptorProto) *descriptorpb.SourceCodeInfo_Location {
 	return s.lookup[e]
 }
 
-func (s sourceCodeInfo) GetEnumValue(e *descriptor.EnumValueDescriptorProto) *descriptor.SourceCodeInfo_Location {
+func (s sourceCodeInfo) GetEnumValue(e *descriptorpb.EnumValueDescriptorProto) *descriptorpb.SourceCodeInfo_Location {
 	return s.lookup[e]
 }
 
-func newSourceCodeInfo(fs []*descriptor.FileDescriptorProto) *sourceCodeInfo {
+func newSourceCodeInfo(fs []*descriptorpb.FileDescriptorProto) *sourceCodeInfo {
 	// For each source location in the provided files
 	// - resolve the (annoyingly) encoded path to its message/field/service/enum/etc definition
 	// - store the source info by its resolved definition
-	lookup := map[proto.Message]*descriptor.SourceCodeInfo_Location{}
+	lookup := map[proto.Message]*descriptorpb.SourceCodeInfo_Location{}
 	for _, f := range fs {
 		for _, loc := range f.GetSourceCodeInfo().GetLocation() {
 			declaration := getDefinitionAtPath(f, loc.Path)
@@ -56,12 +56,12 @@ func newSourceCodeInfo(fs []*descriptor.FileDescriptorProto) *sourceCodeInfo {
 // Resolve a protobuf "file-source path" to its associated definition (eg message/field/enum/etc).
 // Note that some paths don't point to definitions (some reference subcomponents like name, type,
 // field #, etc) and will therefore return nil.
-func getDefinitionAtPath(file *descriptor.FileDescriptorProto, path []int32) proto.Message {
+func getDefinitionAtPath(file *descriptorpb.FileDescriptorProto, path []int32) proto.Message {
 	// The way protobuf encodes "file-source path" is a little opaque/tricky;
 	// this doc describes how it works:
-	//   https://github.com/protocolbuffers/protobuf/blob/master/src/google/protobuf/descriptor.proto#L730
+	//   https://github.com/protocolbuffers/protobuf/blob/master/src/google/protobuf/descriptorpb.proto#L730
 
-	// Starting at the root of the file descriptor, traverse its object graph by following the
+	// Starting at the root of the file descriptorpb, traverse its object graph by following the
 	// specified path (and updating our position/state at each step) until either:
 	// - we reach the definition referenced by the path (and return it)
 	// - we hit a dead end because the path references a grammar element more granular than a
@@ -69,7 +69,7 @@ func getDefinitionAtPath(file *descriptor.FileDescriptorProto, path []int32) pro
 	var pos proto.Message = file
 	for step := 0; step < len(path); step++ {
 		switch p := pos.(type) {
-		case *descriptor.FileDescriptorProto:
+		case *descriptorpb.FileDescriptorProto:
 			switch path[step] {
 			case tag_FileDescriptor_messageType:
 				step++
@@ -81,7 +81,7 @@ func getDefinitionAtPath(file *descriptor.FileDescriptorProto, path []int32) pro
 				return nil // ignore all other types
 			}
 
-		case *descriptor.DescriptorProto:
+		case *descriptorpb.DescriptorProto:
 			switch path[step] {
 			case tag_Descriptor_field:
 				step++
@@ -99,7 +99,7 @@ func getDefinitionAtPath(file *descriptor.FileDescriptorProto, path []int32) pro
 				return nil // ignore all other types
 			}
 
-		case *descriptor.EnumDescriptorProto:
+		case *descriptorpb.EnumDescriptorProto:
 			switch path[step] {
 			case tag_EnumDescriptor_value:
 				step++
